Refactor stealing & destroying bumpers
- Prevents being able to farm extra bumpers off of karma players
- Paves the way for bumper destroy animation later

+++ src/d_clisrv.h
@@ -281,7 +281,7 @@ typedef struct
 	// SRB2kart
 	INT32 kartstuff[NUMKARTSTUFF];
 	tic_t airtime;
-	INT16 bumpers;
+	UINT8 bumpers;
 	INT16 karmadelay;
 	boolean eliminated;
